chore(tests): move accuracy_tests to functional tests as test_inference
