+++ model_api/examples/zsl_visual_prompting/README.md
@@ -31,7 +31,7 @@ Higher thresholds leads to lower mask recall in the final prediction, i.e. low c
 To run the pipeline out-of-the box you can download the test data by running the following command from the repo root:
 
 ```bash
-uv run python tests/accuracy/download_models.py -d data -j tests/accuracy/examples.json -l
+uv run python tests/functional/download_models.py -d data -j tests/functional/examples.json -l
 ```
 
 and then run
