+++ model_api/tests/functional/serving_scope.json
@@ -1,0 +1,6 @@
+[
+  {
+    "name": "otx_models/ssd-card-detection.xml",
+    "type": "DetectionModel"
+  }
+]
