--- model_api/tests/precommit/public_scope.json
@@ -1,38 +1,0 @@
-[
-  {
-    "name": "otx_models/detection_model_with_xai_head.xml",
-    "type": "DetectionModel"
-  },
-  {
-    "name": "otx_models/mlc_mobilenetv3_large_voc.xml",
-    "type": "ClassificationModel"
-  },
-  {
-    "name": "otx_models/Lite-hrnet-18_mod2.xml",
-    "type": "SegmentationModel"
-  },
-  {
-    "name": "otx_models/cls_mobilenetv3_large_cars.onnx",
-    "type": "ClassificationModel"
-  },
-  {
-    "name": "otx_models/tinynet_imagenet.xml",
-    "type": "ClassificationModel"
-  },
-  {
-    "name": "anomalib_models/padim.xml",
-    "type": "AnomalyDetection"
-  },
-  {
-    "name": "anomalib_models/stfpm.xml",
-    "type": "AnomalyDetection"
-  },
-  {
-    "name": "anomalib_models/uflow.xml",
-    "type": "AnomalyDetection"
-  },
-  {
-    "name": "otx_models/ssd-card-detection.xml",
-    "type": "DetectionModel"
-  }
-]
